Use ${JENKINS_REF} variable for /usr/share/jenkins/ref directory
The /usr/share/jenkins/ref directory is an important directory for both,
the Dockerfile for the jenkins image and for derived images. For this
reason it would be more clean and "DRY" to provide and use an
environment variable for this directory instead of using the absolute
path every time.